+++ app/build.gradle
@@ -60,8 +60,8 @@ android {
 
 dependencies {
     coreLibraryDesugaring "com.android.tools:desugar_jdk_libs:2.0.0"
+    implementation project(':core_validation')
 
-    implementation project(":core")
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
