+++ app/src/main/res/layout/fragment_add_story.xml
@@ -9,7 +9,7 @@
 
         <import type="com.don.storyApp.util.StateType" />
 
-        <import type="com.don.core.ValidationType" />
+        <import type="com.don.core_validation.ValidationType" />
 
         <variable
             name="vm"
@@ -39,7 +39,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
 
-            <com.don.core.CustomTextInputLayout
+            <com.don.core_validation.CustomTextInputLayout
                 android:id="@+id/til_description"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -54,7 +54,7 @@
                 app:layout_constraintTop_toBottomOf="@id/iv"
                 app:validationType="@{ValidationType.DEFAULT}">
 
-                <com.don.core.CustomTextInputEditText
+                <com.don.core_validation.CustomTextInputEditText
                     android:id="@+id/ed_add_description"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
@@ -64,7 +64,7 @@
                     android:maxLines="3"
                     android:minLines="3"
                     android:text="@={vm.description}" />
-            </com.don.core.CustomTextInputLayout>
+            </com.don.core_validation.CustomTextInputLayout>
 
             <androidx.appcompat.widget.AppCompatTextView
                 android:id="@+id/tv_take_image"
