+++ core_validation/proguard-rules.pro
