+++ core_validation/src/androidTest/java/com/don/core_validation/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.don.core
+package com.don.core_validation
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
