+++ core_validation/src/main/AndroidManifest.xml
