+++ core_validation/src/main/java/com/don/core_validation/CustomTextInputEditText.kt
@@ -1,4 +1,4 @@
-package com.don.core
+package com.don.core_validation
 
 import android.content.Context
 import android.graphics.Canvas
