+++ core_validation/src/main/java/com/don/core_validation/CustomTextInputLayout.kt
@@ -1,4 +1,4 @@
-package com.don.core
+package com.don.core_validation
 
 import android.content.Context
 import android.text.Editable
