+++ core_validation/src/main/java/com/don/core_validation/ValidationType.kt
@@ -1,4 +1,4 @@
-package com.don.core
+package com.don.core_validation
 
 
 /**
