+++ core_validation/src/main/res/drawable/ic_close.xml
