+++ core_validation/src/main/res/values-en/strings.xml
