+++ core_validation/src/main/res/values-in/strings.xml
