+++ core_validation/src/main/res/values/strings.xml
