+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = "storyapp"
 include ':app'
-include ':core'
+include ':core_validation'
